requirements.txt
line-bot-sdk==1.19.0
python-dotenv==0.19.1
Flask==2.1.1
line-bot-sdk==1.19.0
openai==0.27.0